Move tests to concurrentTest
