+++ tests/integration-tests/src/commonTest/kotlin/test/CacheFlagsTest.kt
@@ -9,15 +9,10 @@ import com.apollographql.apollo.cache.normalized.ApolloStore
 import com.apollographql.apollo.cache.normalized.FetchPolicy
 import com.apollographql.apollo.cache.normalized.api.ApolloCacheHeaders
 import com.apollographql.apollo.cache.normalized.api.CacheHeaders
-import com.apollographql.apollo.cache.normalized.api.MemoryCache
 import com.apollographql.apollo.cache.normalized.api.MemoryCacheFactory
-import com.apollographql.apollo.cache.normalized.api.Record
 import com.apollographql.apollo.cache.normalized.cacheHeaders
 import com.apollographql.apollo.cache.normalized.doNotStore
 import com.apollographql.apollo.cache.normalized.fetchPolicy
-import com.apollographql.apollo.cache.normalized.memoryCacheOnly
-import com.apollographql.apollo.cache.normalized.sql.SqlNormalizedCache
-import com.apollographql.apollo.cache.normalized.sql.SqlNormalizedCacheFactory
 import com.apollographql.apollo.cache.normalized.store
 import com.apollographql.apollo.cache.normalized.storePartialResponses
 import com.apollographql.apollo.exception.CacheMissException
@@ -29,7 +24,6 @@ import com.apollographql.apollo.testing.enqueueTestResponse
 import com.apollographql.apollo.testing.internal.runTest
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.map
-import kotlin.reflect.KClass
 import kotlin.test.Test
 import kotlin.test.assertEquals
 import kotlin.test.assertIs
@@ -134,31 +128,4 @@ class CacheFlagsTest {
         apolloClient.query(query).fetchPolicy(FetchPolicy.CacheOnly).execute().exception
     )
   }
-
-  @Test
-  fun memoryCacheOnlyDoesNotStoreInSqlCache() = runTest {
-    store = ApolloStore(MemoryCacheFactory().chain(SqlNormalizedCacheFactory())).also { it.clearAll() }
-    apolloClient = ApolloClient.Builder().networkTransport(QueueTestNetworkTransport()).store(store).build()
-    val query = HeroNameQuery()
-    apolloClient.enqueueTestResponse(query, HeroNameQuery.Data(HeroNameQuery.Hero("R2-D2")))
-    apolloClient.query(query).memoryCacheOnly(true).execute()
-    val dump: Map<KClass<*>, Map<String, Record>> = store.dump()
-    assertEquals(2, dump[MemoryCache::class]!!.size)
-    assertEquals(0, dump[SqlNormalizedCache::class]!!.size)
-  }
-
-  @Test
-  fun memoryCacheOnlyDoesNotReadFromSqlCache() = runTest {
-    store = ApolloStore(MemoryCacheFactory().chain(SqlNormalizedCacheFactory())).also { it.clearAll() }
-    val query = HeroNameQuery()
-    store.writeOperation(query, HeroNameQuery.Data(HeroNameQuery.Hero("R2-D2")))
-
-    val store2 = ApolloStore(MemoryCacheFactory().chain(SqlNormalizedCacheFactory()))
-    apolloClient = ApolloClient.Builder().serverUrl("unused").store(store2).build()
-    // The record in is in the SQL cache, but we request not to access it
-    assertIs<CacheMissException>(
-        apolloClient.query(query).fetchPolicy(FetchPolicy.CacheOnly).memoryCacheOnly(true).execute().exception
-    )
-  }
-
 }
